personalized_recommendation/requirements.txt
python-dotenv
pymongo
google-generativeai
fastapi
uvicorn
numpy